feat: getServerAuthSession
