+++ src/pages/api/auth/[...nextauth].ts
@@ -1,62 +1,4 @@
-import NextAuth, { NextAuthOptions } from "next-auth";
-import GoogleProvider from "next-auth/providers/google";
-import DiscordProvider from "next-auth/providers/discord";
-import GithubProvider from "next-auth/providers/github";
-import EmailProvider from "next-auth/providers/email";
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import { PrismaClient } from "@prisma/client";
-import { env } from "@env";
-
-const prisma = new PrismaClient();
-
-export const authOptions: NextAuthOptions = {
-  adapter: PrismaAdapter(prisma),
-  providers: [
-    GoogleProvider({
-      clientId: env.GOOGLE_CLIENT_ID,
-      clientSecret: env.GOOGLE_CLIENT_SECRET,
-      allowDangerousEmailAccountLinking: true,
-    }),
-    DiscordProvider({
-      clientId: env.DISCORD_CLIENT_ID,
-      clientSecret: env.DISCORD_CLIENT_SECRET,
-      allowDangerousEmailAccountLinking: true,
-    }),
-    GithubProvider({
-      clientId: env.GITHUB_ID,
-      clientSecret: env.GITHUB_SECRET,
-      allowDangerousEmailAccountLinking: true,
-    }),
-    EmailProvider({
-      server: {
-        host: env.EMAIL_SERVER_HOST,
-        port: Number(env.EMAIL_SERVER_PORT),
-        auth: {
-          user: env.MAILER_USER,
-          pass: env.MAILER_PASSWORD,
-        },
-      },
-      from: '"Leonardo Dias" <leojuriolli@gmail.com>',
-    }),
-  ],
-  theme: {
-    colorScheme: "dark",
-  },
-  pages: {
-    signIn: "/auth/signin",
-    signOut: "/auth/signout",
-    verifyRequest: "/auth/verify-request",
-    error: "/auth/error",
-    newUser: "/auth/new-user",
-  },
-  callbacks: {
-    session: async ({ session, user }) => {
-      session.user.id = user.id;
-      session.user.isAdmin = user.role === "admin";
-
-      return Promise.resolve(session);
-    },
-  },
-};
+import { authOptions } from "@server/utils/auth";
+import NextAuth from "next-auth";
 
 export default NextAuth(authOptions);
