+++ src/pages/posts/liked.tsx
@@ -1,14 +1,14 @@
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import MetaTags from "@components/MetaTags";
-import { authOptions } from "@pages/api/auth/[...nextauth]";
 import type { GetServerSidePropsContext, NextPage } from "next";
-import { getServerSession, User } from "next-auth";
+import type { User } from "next-auth";
 import { trpc } from "@utils/trpc";
 import useOnScreen from "@hooks/useOnScreen";
 import PostCard from "@components/PostCard";
 import ShouldRender from "@components/ShouldRender";
 import EmptyMessage from "@components/EmptyMessage";
 import SearchInput from "@components/SearchInput";
+import { getServerAuthSession } from "@server/utils/auth";
 
 type Props = {
   user: User;
@@ -94,8 +94,11 @@ const UserLikedPage: NextPage<Props> = ({ user }) => {
 
 export default UserLikedPage;
 
-export async function getServerSideProps(context: GetServerSidePropsContext) {
-  const session = await getServerSession(context.req, context.res, authOptions);
+export async function getServerSideProps({
+  req,
+  res,
+}: GetServerSidePropsContext) {
+  const session = await getServerAuthSession({ req, res });
 
   if (!session?.user) {
     return { redirect: { destination: "/" } };
