+++ src/pages/posts/new.tsx
@@ -11,8 +11,6 @@ import Field from "@components/Field";
 import MetaTags from "@components/MetaTags";
 import SelectTags from "@components/SelectTags";
 import type { GetServerSidePropsContext } from "next";
-import { getServerSession } from "next-auth";
-import { authOptions } from "@pages/api/auth/[...nextauth]";
 import Dropzone from "@components/Dropzone";
 import LinkInput from "@components/LinkInput";
 import CreatePoll from "@components/CreatePoll";
@@ -21,6 +19,7 @@ import TextInput from "@components/TextInput";
 import { uploadFileToS3 } from "@utils/aws/uploadFileToS3";
 import { useUploadTagImagesToS3 } from "@hooks/aws/useUploadTagImagesToS3";
 import { parseTagPayload } from "@utils/parseTagPayload";
+import { getServerAuthSession } from "@server/utils/auth";
 
 const CreatePostPage: React.FC = () => {
   const router = useRouter();
@@ -200,8 +199,11 @@ const CreatePostPage: React.FC = () => {
 
 export default CreatePostPage;
 
-export async function getServerSideProps(context: GetServerSidePropsContext) {
-  const session = await getServerSession(context.req, context.res, authOptions);
+export async function getServerSideProps({
+  req,
+  res,
+}: GetServerSidePropsContext) {
+  const session = await getServerAuthSession({ req, res });
 
   // If the user is not logged in, redirect.
   if (!session) {
