+++ src/server/trpc.ts
@@ -1,10 +1,10 @@
 import { type CreateNextContextOptions } from "@trpc/server/adapters/next";
-import { getServerSession, type Session } from "next-auth";
+import type { Session } from "next-auth";
 import { initTRPC, TRPCError } from "@trpc/server";
 import superjson from "superjson";
 import { prisma } from "@utils/prisma";
 import { ZodError } from "zod";
-import { authOptions } from "@pages/api/auth/[...nextauth]";
+import { getServerAuthSession } from "./utils/auth";
 
 type CreateContextOptions = {
   session: Session | null;
@@ -20,7 +20,7 @@ const createInnerTRPCContext = (opts: CreateContextOptions) => {
 export const createTRPCContext = async (opts: CreateNextContextOptions) => {
   const { req, res } = opts;
 
-  const session = await getServerSession(req, res, authOptions);
+  const session = await getServerAuthSession({ req, res });
 
   return createInnerTRPCContext({
     session,
