+++ src/server/utils/auth.ts
@@ -1,0 +1,73 @@
+import type { GetServerSidePropsContext } from "next";
+import { type NextAuthOptions, getServerSession } from "next-auth";
+import GoogleProvider from "next-auth/providers/google";
+import DiscordProvider from "next-auth/providers/discord";
+import GithubProvider from "next-auth/providers/github";
+import EmailProvider from "next-auth/providers/email";
+import { PrismaAdapter } from "@next-auth/prisma-adapter";
+import { PrismaClient } from "@prisma/client";
+import { env } from "@env";
+
+const prisma = new PrismaClient();
+
+export const authOptions: NextAuthOptions = {
+  adapter: PrismaAdapter(prisma),
+  providers: [
+    GoogleProvider({
+      clientId: env.GOOGLE_CLIENT_ID,
+      clientSecret: env.GOOGLE_CLIENT_SECRET,
+      allowDangerousEmailAccountLinking: true,
+    }),
+    DiscordProvider({
+      clientId: env.DISCORD_CLIENT_ID,
+      clientSecret: env.DISCORD_CLIENT_SECRET,
+      allowDangerousEmailAccountLinking: true,
+    }),
+    GithubProvider({
+      clientId: env.GITHUB_ID,
+      clientSecret: env.GITHUB_SECRET,
+      allowDangerousEmailAccountLinking: true,
+    }),
+    EmailProvider({
+      server: {
+        host: env.EMAIL_SERVER_HOST,
+        port: Number(env.EMAIL_SERVER_PORT),
+        auth: {
+          user: env.MAILER_USER,
+          pass: env.MAILER_PASSWORD,
+        },
+      },
+      from: '"Leonardo Dias" <leojuriolli@gmail.com>',
+    }),
+  ],
+  theme: {
+    colorScheme: "dark",
+  },
+  pages: {
+    signIn: "/auth/signin",
+    signOut: "/auth/signout",
+    verifyRequest: "/auth/verify-request",
+    error: "/auth/error",
+    newUser: "/auth/new-user",
+  },
+  callbacks: {
+    session: async ({ session, user }) => {
+      session.user.id = user.id;
+      session.user.isAdmin = user.role === "admin";
+
+      return Promise.resolve(session);
+    },
+  },
+};
+
+/**
+ * Wrapper for `getServerSession` so that you don't need to import the `authOptions` in every file.
+ *
+ * @see https://next-auth.js.org/configuration/nextjs
+ */
+export const getServerAuthSession = (ctx: {
+  req: GetServerSidePropsContext["req"];
+  res: GetServerSidePropsContext["res"];
+}) => {
+  return getServerSession(ctx.req, ctx.res, authOptions);
+};
